chore: refactor init and deinit logic to Window
